services/restapi/requirements.txt
flask_marshmallow==0.14.0
Flask==1.1.2
Flask_Migrate==2.5.3
Flask_SQLAlchemy==2.4.4
flask_restx==0.2.0
Flask_Login==0.5.0
itsdangerous==1.1.0
Werkzeug==0.16.0
celery==5.0.5
flask_bootstrap==3.3.7.1
flask_mail==0.9.1
flask_wtf==0.14.3
python-decouple==3.4
python-dotenv==0.15.0
WTForms==2.3.3
marshmallow-sqlalchemy
psycopg2
healthcheck
Flask-JWT-Extended
PyJWT
passlib
Flask-HTTPAuth